JE2-77 [FE] ADDITIONAL Hooks for ClientForm and ProductForm moved to the right places.
