+++ frontend/src/features/clients/components/ClientForm.tsx
@@ -1,6 +1,6 @@
 import Grid from '@mui/material/Grid2'
 import { Button, CircularProgress, TextField, Typography } from '@mui/material'
-import { useClientForm } from '../../../hooks/useClientForm.ts'
+import { useClientForm } from '../hooks/useClientForm.ts'
 import { useParams } from 'react-router-dom'
 import { Client } from '../../../types'
