+++ frontend/src/features/clients/hooks/useClientForm.ts
@@ -1,10 +1,10 @@
 import { ChangeEvent, FormEvent, useEffect, useState } from 'react'
 import { toast } from 'react-toastify'
-import { ClientMutation } from '../types'
-import { useAppDispatch, useAppSelector } from '../app/hooks.ts'
-import { selectClient, selectClientError, selectLoadingAddClient } from '../store/slices/clientSlice.ts'
-import { addClient, fetchClientById, updateClient } from '../store/thunks/clientThunk.ts'
-import { emailRegex, phoneNumberRegex, initialClientState } from '../constants.ts'
+import { ClientMutation } from '../../../types'
+import { useAppDispatch, useAppSelector } from '../../../app/hooks.ts'
+import { selectClient, selectClientError, selectLoadingAddClient } from '../../../store/slices/clientSlice.ts'
+import { addClient, fetchClientById, updateClient } from '../../../store/thunks/clientThunk.ts'
+import { emailRegex, phoneNumberRegex, initialClientState } from '../../../constants.ts'
 
 const requiredFields: (keyof ClientMutation)[] = ['name', 'email', 'phone_number', 'inn']
