+++ frontend/src/features/products/components/ProductForm.tsx
@@ -1,6 +1,6 @@
 import Grid from '@mui/material/Grid2'
 import { Button, CircularProgress, TextField, Typography, Box, Autocomplete } from '@mui/material'
-import useProductForm from '../../../hooks/useProductForm.ts'
+import useProductForm from '../hooks/useProductForm.ts'
 
 const ProductForm = () => {
